feat: add loader
