+++ src/components/Loader/Loader.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Spinner } from 'react-bootstrap';
+import './loader.scss';
+
+const Loader = () => (
+  <div className="loader__container">
+    <div className="loader">
+      <Spinner animation="grow  text-info" role="status">
+        <span className="sr-only">Загрузка...</span>
+      </Spinner>
+    </div>
+  </div>
+);
+
+export default Loader;
