+++ src/components/Loader/loader.scss
@@ -1,0 +1,10 @@
+.loader__container{
+    height: 100%;
+    width: 100%;
+    position: fixed;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #0000008c;
+    z-index: 995;
+}
