+++ src/components/sharingan/Sharingan.jsx
@@ -1,22 +1,13 @@
 import React, { useState } from 'react';
-import getImages from '../../services/backgroundImage.service';
+import setBodyImage from '../../utils/setBodyImage';
 import './sharingan.scss';
 
 const Sharingan = () => {
   const [isLoading, setIsLoading] = useState(false);
 
-  const changeBackgroundImage = async () => {
+  const changeBackgroundImage = () => {
     setIsLoading(true);
-    const body = document.querySelector('body');
-    const src = getImages();
-    const img = new Image();
-    img.src = src;
-    img.onload = () => {
-      if (src) {
-        body.style.backgroundImage = `url(${src})`;
-      }
-      setTimeout(() => setIsLoading(false), 500);
-    };
+    setBodyImage().finally(() => setTimeout(() => setIsLoading(false), 500));
   };
   return (
     <div className={`circle circle-in ${isLoading ? 'circle-load' : ''}`} onClick={changeBackgroundImage}>
