+++ src/main.jsx
@@ -4,11 +4,8 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import './assets/styles/main.scss';
 import './assets/styles/icons.min.css';
-import animation from './utils/backgroundAnimation';
 
 ReactDOM.render(
   <App />,
   document.getElementById('root'),
 );
-
-animation();
